truthshield_ai/requirements.txt
streamlit
pyrebase4
langchain
langchain-community
langchain-huggingface
faiss-cpu
sentence-transformers
transformers
torch
